Using the more-updated circonus labs nad
and changing loads of stuff to make it all work, removing now-outdated
junk in the process.  Also adding the concept of index rebuilding, but
not sure if it really matters anymore (???)  Nice to have, I suppose.
In the end, it looks like @clintoncwolfe has his cookbook in a much
better place.  Yay!