Making sure node is available in PATH if already installed
It appears that in some cases, at least on Joyent, node is installed but
in a special place, and attempts to install node otherwise break.
Therefore, if node can be found on the system using `find`, it is
symlinked into a location that can be found.  Then, if it still cannot be
found (i.e. `find` couldn't find it), it is installed as a package