Only installing node from source on Linux
On linux, installing with apt results in an old version, but installing
from source works.  On SmartOS, installing from source would require
epic munging of the nad install with yet another patch to update the smf
manifest to include a different `LD_LIBRARY_PATH`.  Ugh.  So instead of
doing that, we're ensuring the lastest version with the `package`
provider.  If used on a SmartOS machine with nad already working, this
should have no effect